officially sanction the slightly strange ways that scuttlebot uses muxrpc - then a line in the sand can be drawn
